fix:　起動できない問題の修正
