+++ app/index.tsx
@@ -1,1 +1,44 @@
-export { default } from '@/screens/index';
+import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { useRouter } from 'expo-router';
+
+export default function Index() {
+  const router = useRouter();
+
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Hello World - Test Screen</Text>
+
+      <TouchableOpacity
+        style={styles.button}
+        onPress={() => router.push('/user')}
+      >
+        <Text style={styles.buttonText}>Go to User Page</Text>
+      </TouchableOpacity>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#fff',
+    gap: 20,
+  },
+  title: {
+    fontSize: 18,
+    fontWeight: '600',
+  },
+  button: {
+    backgroundColor: '#007AFF',
+    paddingHorizontal: 24,
+    paddingVertical: 12,
+    borderRadius: 8,
+  },
+  buttonText: {
+    color: '#fff',
+    fontSize: 16,
+    fontWeight: '600',
+  },
+});
