--- screens/(tabs)/explore-styles.ts
@@ -1,20 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-  headerImage: {
-    color: '#808080',
-    bottom: -90,
-    left: -35,
-    position: 'absolute',
-  },
-  titleContainer: {
-    flexDirection: 'row',
-    gap: 8,
-  },
-  centeredImage: {
-    alignSelf: 'center',
-  },
-  spaceMonoFont: {
-    fontFamily: 'SpaceMono',
-  },
-});
