+++ screens/user/index.tsx
@@ -1,20 +1,74 @@
-import React, { useState } from 'react';
-import { View, Text, TouchableOpacity, Modal } from 'react-native';
-import { Stack } from 'expo-router';
-import MaterialIcons from '@expo/vector-icons/MaterialIcons';
-import { UserHomeLayout } from '@/components/layouts/UserHomeLayout/UserHomeLayout';
-import { BottomNavigation } from '@/components/layouts/BottomNavigation/BottomNavigation';
-import { styles } from './styles';
+import { BottomNavigation } from "@/components/layouts/BottomNavigation/BottomNavigation";
+import { UserHomeLayout } from "@/components/layouts/UserHomeLayout/UserHomeLayout";
+import MaterialIcons from "@expo/vector-icons/MaterialIcons";
+import { Stack, useRouter } from "expo-router";
+import React, { useState } from "react";
+import { Modal, Text, TouchableOpacity, View } from "react-native";
+import { styles } from "./styles";
 
 export default function UserHomeScreen() {
   const [showHelpModal, setShowHelpModal] = useState(false);
+  const router = useRouter();
+
+  // Mock data - 将来的にはAPIから取得
+  const fatigueLevel = 45; // 疲労度 (0-100)
+  const fatigueStatus = "注意"; // 良好 | 注意 | 警告
+
+  const urgentNotifications = [
+    {
+      id: "1",
+      title: "水分補給の時間です",
+      description: "前回の水分補給から2時間が経過しました",
+      time: "5分前",
+      color: "#FFA726",
+    },
+    {
+      id: "2",
+      title: "休憩のおすすめ",
+      description: "疲労度が上昇しています。15分程度の休憩をお勧めします",
+      time: "15分前",
+      color: "#42A5F5",
+    },
+  ];
+
+  const recentActivities = [
+    "午前の散歩を完了しました",
+    "水分補給リマインダー",
+    "朝のバイタル測定完了",
+  ];
+
+  // 疲労度に応じた背景色とボーダー色を取得
+  const getFatigueColors = (level: number) => {
+    if (level < 30) {
+      return { bg: "#E8F5E9", border: "#4CAF50", progress: "#4CAF50" }; // 緑系
+    } else if (level < 60) {
+      return { bg: "#FFF8E1", border: "#FFA726", progress: "#FFA726" }; // 黄色/オレンジ系
+    } else {
+      return { bg: "#FFEBEE", border: "#EF5350", progress: "#EF5350" }; // 赤系
+    }
+  };
+
+  const fatigueColors = getFatigueColors(fatigueLevel);
 
   return (
     <>
       <Stack.Screen options={{ headerShown: false }} />
       <View style={styles.container}>
         <UserHomeLayout>
-          {/* User Profile Card */}
+          {/* Page Title */}
+          <View style={styles.pageHeader}>
+            <Text style={styles.pageTitle}>ホーム</Text>
+            <TouchableOpacity
+              style={styles.switchViewButton}
+              onPress={() => router.push("/helper")}
+            >
+              <Text style={styles.switchViewButtonText}>
+                介助者ビューに切替
+              </Text>
+            </TouchableOpacity>
+          </View>
+
+          {/* User Profile Card - Compact */}
           <View style={styles.profileCard}>
             <View style={styles.profileLeft}>
               <View style={styles.avatarCircle}>
@@ -25,7 +79,10 @@ export default function UserHomeScreen() {
                 <Text style={styles.profileStatus}>状態: 良好</Text>
               </View>
             </View>
-            <TouchableOpacity style={styles.settingsButton}>
+            <TouchableOpacity
+              style={styles.settingsButton}
+              onPress={() => router.push("/user/settings")}
+            >
               <Text style={styles.settingsButtonText}>設定</Text>
             </TouchableOpacity>
           </View>
@@ -35,69 +92,98 @@ export default function UserHomeScreen() {
             style={styles.helpButton}
             onPress={() => setShowHelpModal(true)}
           >
-            <View style={styles.helpButtonContent}>
-              <View style={styles.helpIcon}>
-                <Text style={styles.helpIconText}>⚠</Text>
-              </View>
-              <Text style={styles.helpButtonText}>ヘルプ要請</Text>
-            </View>
+            <MaterialIcons name="error-outline" size={24} color="#FFFFFF" />
+            <Text style={styles.helpButtonText}>ヘルプ要請</Text>
           </TouchableOpacity>
 
-          {/* Vital Data Section */}
+          {/* Fatigue Level Card */}
+          <View
+            style={[
+              styles.fatigueCard,
+              {
+                backgroundColor: fatigueColors.bg,
+                borderColor: fatigueColors.border,
+              },
+            ]}
+          >
+            <View style={styles.fatigueHeader}>
+              <View style={styles.fatigueTitle}>
+                <MaterialIcons name="show-chart" size={20} color="#333333" />
+                <Text style={styles.fatigueTitleText}>現在の疲労度</Text>
+              </View>
+              <View style={styles.fatigueStatusBadge}>
+                <Text style={styles.fatigueStatusText}>{fatigueStatus}</Text>
+              </View>
+            </View>
+
+            <View style={styles.fatigueContent}>
+              <Text style={styles.fatigueValue}>{fatigueLevel}</Text>
+              <Text style={styles.fatigueUnit}>%</Text>
+            </View>
+
+            {/* Progress Bar */}
+            <View style={styles.fatigueProgressContainer}>
+              <View
+                style={[
+                  styles.fatigueProgressBar,
+                  {
+                    backgroundColor: fatigueColors.progress,
+                    width: `${fatigueLevel}%`,
+                  },
+                ]}
+              />
+            </View>
+
+            <Text style={styles.fatigueDescription}>
+              センサーから取得した疲労度データです。休憩が必要な場合はお知らせします。
+            </Text>
+          </View>
+
+          {/* Urgent Notifications Section */}
           <View style={styles.section}>
-            <Text style={styles.sectionTitle}>バイタルデータ</Text>
-            <View style={styles.vitalGrid}>
-              {/* Heart Rate */}
-              <View style={styles.vitalCard}>
-                <View style={styles.vitalHeader}>
-                  <MaterialIcons name="favorite" size={18} color="#FF6B6B" />
-                  <Text style={styles.vitalLabel}>心拍数</Text>
-                </View>
-                <Text style={styles.vitalValue}>72 <Text style={styles.vitalUnit}>bpm</Text></Text>
-                <View style={[styles.vitalBar, { backgroundColor: '#FFE5E5' }]}>
-                  <View style={[styles.vitalBarFill, { width: '70%', backgroundColor: '#FF6B6B' }]} />
-                </View>
-                <Text style={styles.vitalStatus}>正常</Text>
-              </View>
+            <View style={styles.sectionHeader}>
+              <MaterialIcons
+                name="notifications-active"
+                size={20}
+                color="#FF6B6B"
+              />
+              <Text style={styles.sectionTitle}>要対応</Text>
+            </View>
 
-              {/* Health Index */}
-              <View style={styles.vitalCard}>
-                <View style={styles.vitalHeader}>
-                  <MaterialIcons name="trending-up" size={18} color="#20C9A6" />
-                  <Text style={styles.vitalLabel}>健康指標</Text>
+            <View style={styles.urgentNotificationsList}>
+              {urgentNotifications.map((notification) => (
+                <View
+                  key={notification.id}
+                  style={styles.urgentNotificationItem}
+                >
+                  <View
+                    style={[
+                      styles.urgentNotificationIndicator,
+                      { backgroundColor: notification.color },
+                    ]}
+                  />
+                  <View style={styles.urgentNotificationContent}>
+                    <View style={styles.urgentNotificationTop}>
+                      <Text style={styles.urgentNotificationTitle}>
+                        {notification.title}
+                      </Text>
+                      <View style={styles.urgentNotificationTime}>
+                        <MaterialIcons
+                          name="access-time"
+                          size={14}
+                          color="#999999"
+                        />
+                        <Text style={styles.urgentNotificationTimeText}>
+                          {notification.time}
+                        </Text>
+                      </View>
+                    </View>
+                    <Text style={styles.urgentNotificationDescription}>
+                      {notification.description}
+                    </Text>
+                  </View>
                 </View>
-                <Text style={styles.vitalValue}>85 <Text style={styles.vitalUnit}>%</Text></Text>
-                <View style={[styles.vitalBar, { backgroundColor: '#E0F7F7' }]}>
-                  <View style={[styles.vitalBarFill, { width: '85%', backgroundColor: '#20C9A6' }]} />
-                </View>
-                <Text style={styles.vitalStatus}>良好</Text>
-              </View>
-
-              {/* Activity Level */}
-              <View style={styles.vitalCard}>
-                <View style={styles.vitalHeader}>
-                  <MaterialIcons name="flash-on" size={18} color="#9B6CFF" />
-                  <Text style={styles.vitalLabel}>活動レベル</Text>
-                </View>
-                <Text style={styles.vitalValue}>320 <Text style={styles.vitalUnit}>kcal</Text></Text>
-                <View style={[styles.vitalBar, { backgroundColor: '#F0E6FF' }]}>
-                  <View style={[styles.vitalBarFill, { width: '60%', backgroundColor: '#9B6CFF' }]} />
-                </View>
-                <Text style={styles.vitalStatus}>中程度</Text>
-              </View>
-
-              {/* Water Intake */}
-              <View style={styles.vitalCard}>
-                <View style={styles.vitalHeader}>
-                  <MaterialIcons name="water-drop" size={18} color="#2196F3" />
-                  <Text style={styles.vitalLabel}>水分補給</Text>
-                </View>
-                <Text style={styles.vitalValue}>1.2 <Text style={styles.vitalUnit}>L</Text></Text>
-                <View style={[styles.vitalBar, { backgroundColor: '#E3F2FD' }]}>
-                  <View style={[styles.vitalBarFill, { width: '60%', backgroundColor: '#2196F3' }]} />
-                </View>
-                <Text style={styles.vitalStatus}>目標の60%</Text>
-              </View>
+              ))}
             </View>
           </View>
 
@@ -105,7 +191,12 @@ export default function UserHomeScreen() {
           <View style={styles.section}>
             <Text style={styles.sectionTitle}>最近の活動</Text>
             <View style={styles.activityList}>
-              {/* Activity items would go here */}
+              {recentActivities.map((activity, index) => (
+                <View key={index} style={styles.activityItem}>
+                  <View style={styles.activityBullet} />
+                  <Text style={styles.activityText}>{activity}</Text>
+                </View>
+              ))}
             </View>
           </View>
         </UserHomeLayout>
@@ -133,7 +224,7 @@ function HelpRequestModal({ visible, onClose }: HelpRequestModalProps) {
   const [includeLocation, setIncludeLocation] = useState(true);
 
   const handleSendHelp = () => {
-    console.log('Send help request with location:', includeLocation);
+    console.log("Send help request with location:", includeLocation);
     onClose();
   };
 
@@ -171,8 +262,17 @@ function HelpRequestModal({ visible, onClose }: HelpRequestModalProps) {
               <Text style={styles.modalSectionTitle}>送信先</Text>
               <View style={styles.recipientsList}>
                 <View style={styles.recipientItem}>
-                  <View style={[styles.recipientAvatar, { backgroundColor: '#FFE5E5' }]}>
-                    <Text style={[styles.recipientAvatarText, { color: '#FF6B6B' }]}>山</Text>
+                  <View
+                    style={[
+                      styles.recipientAvatar,
+                      { backgroundColor: "#FFE5E5" },
+                    ]}
+                  >
+                    <Text
+                      style={[styles.recipientAvatarText, { color: "#FF6B6B" }]}
+                    >
+                      山
+                    </Text>
                   </View>
                   <View style={styles.recipientInfo}>
                     <Text style={styles.recipientName}>山田花子</Text>
@@ -180,8 +280,17 @@ function HelpRequestModal({ visible, onClose }: HelpRequestModalProps) {
                   </View>
                 </View>
                 <View style={styles.recipientItem}>
-                  <View style={[styles.recipientAvatar, { backgroundColor: '#FFE5E5' }]}>
-                    <Text style={[styles.recipientAvatarText, { color: '#FF6B6B' }]}>佐</Text>
+                  <View
+                    style={[
+                      styles.recipientAvatar,
+                      { backgroundColor: "#FFE5E5" },
+                    ]}
+                  >
+                    <Text
+                      style={[styles.recipientAvatarText, { color: "#FF6B6B" }]}
+                    >
+                      佐
+                    </Text>
                   </View>
                   <View style={styles.recipientInfo}>
                     <Text style={styles.recipientName}>佐藤健太</Text>
@@ -223,10 +332,12 @@ function HelpRequestModal({ visible, onClose }: HelpRequestModalProps) {
                 <Text style={styles.locationSubtitle}>現在地: 自宅</Text>
               </View>
               <View style={styles.checkboxContainer}>
-                <View style={[
-                  styles.checkbox,
-                  includeLocation && styles.checkboxChecked
-                ]}>
+                <View
+                  style={[
+                    styles.checkbox,
+                    includeLocation && styles.checkboxChecked,
+                  ]}
+                >
                   {includeLocation && (
                     <Text style={styles.checkboxIcon}>✓</Text>
                   )}
@@ -237,10 +348,7 @@ function HelpRequestModal({ visible, onClose }: HelpRequestModalProps) {
 
           {/* Modal Footer */}
           <View style={styles.modalFooter}>
-            <TouchableOpacity
-              style={styles.cancelButton}
-              onPress={onClose}
-            >
+            <TouchableOpacity style={styles.cancelButton} onPress={onClose}>
               <Text style={styles.cancelButtonText}>キャンセル</Text>
             </TouchableOpacity>
             <TouchableOpacity
